LPS-58289 Move reindexing code to its proper indexer
